Fix for disappearing files
If you have a good Internet connection, uploading multiple files at the same time leads to the loss of some. That is expressed in the non-saving of data in the session variable when competing requests.. Strange bug but this is only way to fix. Also this is fix for issue https://github.com/Jako/AjaxUpload/issues/18